// generated/src/app/api-models/delegation-service-codes.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class DelegationServiceCodes {
  delegationServiceCode: string;
  delegationServiceDesc: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}